feat(parse): add support for parsing struct and slice fields
This commit introduces the ability to parse struct and slice fieldswithin the gen/parse module. This is a significant enhancement that
expands the functionality of the parser, allowing for more complex
data structures to be processed.

In addition, the commit includes the parsing of selector expressions
to improve type representation in generated output. This furtherbolsters the flexibility and usefulness of the parsing capabilities.

BREAKING CHANGE: The addition of new field types in the parse logic
may affect existing consumers of the parse API that do not expect
struct or slice fields to be present in the parsed output.